refactor scheduler commands page
