+++ src/components/etke.cc/scheduler/hooks/useRecurringCommands.tsx
@@ -1,0 +1,15 @@
+import { useDataProvider } from "react-admin";
+
+import { useQuery } from "@tanstack/react-query";
+import { useAppContext } from "../../../../Context";
+
+export const useRecurringCommands = () => {
+  const { etkeccAdmin } = useAppContext();
+  const dataProvider = useDataProvider();
+  const { data, isLoading, error } = useQuery({
+    queryKey: ["recurringCommands"],
+    queryFn: () => dataProvider.getRecurringCommands(etkeccAdmin),
+  });
+
+  return { data, isLoading, error };
+};
